# backend/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  my-web-app_com_smart-logistic-hub_internal_inventory_dto.CreateInventoryRequest:
    properties:
      available_qty:
        type: integer
      damaged_qty:
        type: integer
      hold_qty:
        type: integer
      product_id:
        type: integer
      reserved_qty:
        type: integer
      updated_by:
        type: string
      warehouse_id:
        type: integer
    required:
    - product_id
    - warehouse_id
    type: object
  my-web-app_com_smart-logistic-hub_internal_inventory_dto.InventoryResponse:
    properties:
      available_qty:
        type: integer
      created_at:
        type: string
      damaged_qty:
        type: integer
      hold_qty:
        type: integer
      id:
        type: integer
      product_id:
        type: integer
      reserved_qty:
        type: integer
      updated_at:
        type: string
      updated_by:
        type: string
      warehouse_id:
        type: integer
    type: object
  my-web-app_com_smart-logistic-hub_internal_inventory_dto.PaginatedResponse:
    properties:
      items:
        items:
          $ref: '#/definitions/my-web-app_com_smart-logistic-hub_internal_inventory_dto.InventoryResponse'
        type: array
      limit:
        type: integer
      skip:
        type: integer
      total:
        type: integer
    type: object
  my-web-app_com_smart-logistic-hub_internal_inventory_dto.UpdateInventoryRequest:
    properties:
      available_qty:
        type: integer
      damaged_qty:
        type: integer
      hold_qty:
        type: integer
      reason:
        type: string
      reserved_qty:
        type: integer
      updated_by:
        type: string
    type: object
  my-web-app_com_smart-logistic-hub_internal_order_dto.CreateOrderRequest:
    properties:
      assigned_driver_id:
        type: integer
      items:
        items:
          $ref: '#/definitions/my-web-app_com_smart-logistic-hub_internal_order_dto.OrderItemRequest'
        type: array
      order_code:
        type: string
      receiver_address:
        type: string
      receiver_district:
        type: string
      receiver_name:
        type: string
      receiver_phone:
        type: string
      receiver_postal_code:
        type: string
      receiver_province:
        type: string
      receiver_ward:
        type: string
      sender_address:
        type: string
      sender_district:
        type: string
      sender_name:
        type: string
      sender_phone:
        type: string
      sender_postal_code:
        type: string
      sender_province:
        type: string
      sender_ward:
        type: string
      status:
        enum:
        - PENDING
        - RESERVED
        - PICKING
        - PACKING
        - SORTING
        - SHIPPING
        - COMPLETED
        - PICKING_UP
        type: string
      warehouse_id:
        minimum: 1
        type: integer
    required:
    - order_code
    - receiver_address
    - receiver_name
    - receiver_phone
    - sender_address
    - sender_name
    - sender_phone
    - warehouse_id
    type: object
  my-web-app_com_smart-logistic-hub_internal_order_dto.OrderItemRequest:
    properties:
      product_id:
        type: integer
      product_name:
        type: string
      quantity:
        minimum: 1
        type: integer
      weight_gram:
        type: integer
    required:
    - quantity
    type: object
  my-web-app_com_smart-logistic-hub_internal_order_dto.OrderItemResponse:
    properties:
      id:
        type: integer
      order_id:
        type: integer
      product_id:
        type: integer
      product_name:
        type: string
      quantity:
        type: integer
      weight_gram:
        type: integer
    type: object
  my-web-app_com_smart-logistic-hub_internal_order_dto.OrderResponse:
    properties:
      assigned_driver_id:
        type: integer
      created_at:
        type: string
      created_by:
        type: string
      id:
        type: integer
      items:
        items:
          $ref: '#/definitions/my-web-app_com_smart-logistic-hub_internal_order_dto.OrderItemResponse'
        type: array
      order_code:
        type: string
      receiver_address:
        type: string
      receiver_district:
        type: string
      receiver_name:
        type: string
      receiver_phone:
        type: string
      receiver_postal_code:
        type: string
      receiver_province:
        type: string
      receiver_ward:
        type: string
      sender_address:
        type: string
      sender_district:
        type: string
      sender_name:
        type: string
      sender_phone:
        type: string
      sender_postal_code:
        type: string
      sender_province:
        type: string
      sender_ward:
        type: string
      status:
        type: string
      updated_at:
        type: string
    type: object
  my-web-app_com_smart-logistic-hub_internal_order_dto.PaginatedResponse:
    properties:
      items:
        items:
          $ref: '#/definitions/my-web-app_com_smart-logistic-hub_internal_order_dto.OrderResponse'
        type: array
      limit:
        type: integer
      skip:
        type: integer
      total:
        type: integer
    type: object
  my-web-app_com_smart-logistic-hub_internal_order_dto.UpdateOrderRequest:
    properties:
      assigned_driver_id:
        type: integer
      order_code:
        type: string
      receiver_address:
        type: string
      receiver_district:
        type: string
      receiver_name:
        type: string
      receiver_phone:
        type: string
      receiver_postal_code:
        type: string
      receiver_province:
        type: string
      receiver_ward:
        type: string
      sender_address:
        type: string
      sender_district:
        type: string
      sender_name:
        type: string
      sender_phone:
        type: string
      sender_postal_code:
        type: string
      sender_province:
        type: string
      sender_ward:
        type: string
      status:
        enum:
        - PENDING
        - RESERVED
        - PICKING
        - PACKING
        - SORTING
        - SHIPPING
        - COMPLETED
        - PICKING_UP
        type: string
    type: object
...
host: localhost:8000
info:
  contact: {}
  description: Smart Logistics Hub backend API. Authentication uses the "Authorization"
    header with format "Bearer &lt;token&gt;" (JWT issued by Keycloak).
  title: Smart Logistics Hub API
  version: 1.0.0
paths:
...
  /inventory:
    get:
      parameters:
      - default: 0
        description: Number of items to skip
        in: query
        name: skip
        type: integer
      - default: 20
        description: Max items per page
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/my-web-app_com_smart-logistic-hub_internal_inventory_dto.PaginatedResponse'
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: List inventory records
      tags:
      - Inventory
    post:
      consumes:
      - application/json
      parameters:
      - description: Inventory payload
        in: body
        name: inventory
        required: true
        schema:
          $ref: '#/definitions/my-web-app_com_smart-logistic-hub_internal_inventory_dto.CreateInventoryRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/my-web-app_com_smart-logistic-hub_internal_inventory_dto.InventoryResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Create an inventory record
      tags:
      - Inventory
  /inventory/{id}:
    delete:
      parameters:
      - description: Inventory ID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Delete an inventory record
      tags:
      - Inventory
    get:
      parameters:
      - description: Inventory ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/my-web-app_com_smart-logistic-hub_internal_inventory_dto.InventoryResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Get an inventory record by ID
      tags:
      - Inventory
    patch:
      consumes:
      - application/json
      parameters:
      - description: Inventory ID
        in: path
        name: id
        required: true
        type: integer
      - description: Inventory update
        in: body
        name: inventory
        required: true
        schema:
          $ref: '#/definitions/my-web-app_com_smart-logistic-hub_internal_inventory_dto.UpdateInventoryRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/my-web-app_com_smart-logistic-hub_internal_inventory_dto.InventoryResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Update an inventory record
      tags:
      - Inventory
  /orders:
    get:
      parameters:
      - default: 0
        description: Number of items to skip
        in: query
        name: skip
        type: integer
      - default: 10
        description: Max items per page
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/my-web-app_com_smart-logistic-hub_internal_order_dto.PaginatedResponse'
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: List orders
      tags:
      - Orders
    post:
      consumes:
      - application/json
      parameters:
      - description: Order payload
        in: body
        name: order
        required: true
        schema:
          $ref: '#/definitions/my-web-app_com_smart-logistic-hub_internal_order_dto.CreateOrderRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/my-web-app_com_smart-logistic-hub_internal_order_dto.OrderResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Create an order
      tags:
      - Orders
  /orders/{id}:
    delete:
      parameters:
      - description: Order ID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Delete an order
      tags:
      - Orders
    get:
      parameters:
      - description: Order ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/my-web-app_com_smart-logistic-hub_internal_order_dto.OrderResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Get an order by ID
      tags:
      - Orders
    patch:
      consumes:
      - application/json
      parameters:
      - description: Order ID
        in: path
        name: id
        required: true
        type: integer
      - description: Order update
        in: body
        name: order
        required: true
        schema:
          $ref: '#/definitions/my-web-app_com_smart-logistic-hub_internal_order_dto.UpdateOrderRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/my-web-app_com_smart-logistic-hub_internal_order_dto.OrderResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Update an order
      tags:
      - Orders
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"